ss: windows build adaptations
Windows compiler finds various non-bug things to complain about when
building with SS and other options, fix them up

+++ lib/misc/lws-struct-lejp.c
@@ -675,7 +675,7 @@ lws_struct_json_serialize(lws_struct_serialize_t *js, uint8_t *buf,
 			 * in "used".
 			 */
 
-			lws_json_purify((char *)buf, q, len, &used);
+			lws_json_purify((char *)buf, q, (int)len, &used);
 			m = strlen((const char *)buf);
 			buf += m;
 			len -= m;
